+++ src/onegov.org/onegov/org/layout.py
@@ -381,6 +381,10 @@ class Layout(ChameleonLayout):
     def get_user_color(self, username):
         return utils.get_user_color(username)
 
+    def get_user_title(self, username):
+        user = UserCollection(self.request.session).by_username(username)
+        return user and user.title or username
+
     def to_timezone(self, date, timezone):
         return to_timezone(date, timezone)