+++ src/onegov.org/onegov/org/models/__init__.py
@@ -22,6 +22,7 @@ from onegov.org.models.legacy_file import (
 )
 from onegov.org.models.message import DirectoryMessage
 from onegov.org.models.message import EventMessage
+from onegov.org.models.message import FileSignatureMessage
 from onegov.org.models.message import PaymentMessage
 from onegov.org.models.message import ReservationMessage
 from onegov.org.models.message import SubmissionMessage
@@ -71,6 +72,7 @@ __all__ = [
     'ExportCollection',
     'ExtendedDirectory',
     'ExtendedDirectoryEntry',
+    'FileSignatureMessage',
     'FormPersonMove',
     'GeneralFile',
     'GeneralFileCollection',