+++ src/onegov.org/onegov/org/tests/test_views.py
@@ -6,6 +6,7 @@ import re
 import requests_mock
 import textwrap
 import transaction
+import vcr
 
 from base64 import b64decode, b64encode
 from datetime import datetime, date, timedelta
@@ -14,20 +15,24 @@ from libres.modules.errors import AffectedReservationError
 from lxml.html import document_fromstring
 from onegov.core.custom import json
 from onegov.core.utils import Bunch
-from onegov.form import FormCollection, FormSubmission
+from onegov.core.utils import module_path
 from onegov.directory import DirectoryEntry
+from onegov.file import FileCollection
+from onegov.form import FormCollection, FormSubmission
 from onegov.gis import Coordinates
 from onegov.newsletter import RecipientCollection, NewsletterCollection
 from onegov.page import PageCollection
 from onegov.pay import PaymentProviderCollection
 from onegov.people import Person
 from onegov.reservation import ResourceCollection, Reservation
-from onegov_testing import utils
 from onegov.ticket import TicketCollection
 from onegov.user import UserCollection
-from sedate import replace_timezone
+from onegov_testing import utils
 from purl import URL
+from sedate import replace_timezone
+from unittest.mock import patch
 from webtest import Upload
+from yubico_client import Yubico
 
 
 def encode_map_value(dictionary):
@@ -4362,3 +4367,111 @@ def test_markdown_in_directories(client):
     page.form.submit()
 
     assert "<li>Soccer rules" in client.get('/directories/clubs/soccer-club')
+
+
+def test_search_signed_files(client_with_es):
+    client = client_with_es
+    client.login_admin()
+
+    path = module_path('onegov.org', 'tests/fixtures/sample.pdf')
+    with open(path, 'rb') as f:
+        page = client.get('/files')
+        page.form['file'] = Upload('Sample.pdf', f.read(), 'application/pdf')
+        page.form.submit()
+
+    client.app.es_indexer.process()
+    client.app.es_client.indices.refresh(index='_all')
+
+    assert 'Sample' in client.get('/search?q=Adobe')
+    assert 'Sample' not in client.spawn().get('/search?q=Adobe')
+
+    transaction.begin()
+    pdf = FileCollection(client.app.session()).query().one()
+    pdf.signed = True
+    transaction.commit()
+
+    client.app.es_indexer.process()
+    client.app.es_client.indices.refresh(index='_all')
+
+    assert 'Sample' in client.get('/search?q=Adobe')
+    assert 'Sample' in client.spawn().get('/search?q=Adobe')
+
+
+def test_sign_document(client):
+    client.login_admin()
+
+    path = module_path('onegov.org', 'tests/fixtures/sample.pdf')
+    with open(path, 'rb') as f:
+        page = client.get('/files')
+        page.form['file'] = Upload('Sample.pdf', f.read(), 'application/pdf')
+        page.form.submit()
+
+    pdf = FileCollection(client.app.session()).query().one()
+
+    # signatures are only used if yubikeys are used
+    page = client.get(f'/storage/{pdf.id}/details')
+    assert 'signature' not in page
+
+    client.app.enable_yubikey = True
+
+    page = client.get(f'/storage/{pdf.id}/details')
+    assert 'signature' in page
+
+    # signing only works if the given user has a yubikey setup
+    def sign(client, page, token):
+        rex = r"'(http://\w+/\w+/\w+/sign?[^']+)'"
+        url = re.search(rex, str(page)).group(1)
+        return client.post(url, {'token': token})
+
+    assert "Bitte geben Sie Ihren Yubikey ein" in sign(client, page, '')
+    assert "nicht mit einem Yubikey verknüpft" in sign(client, page, 'foobar')
+
+    # not just any yubikey either, it has to be the one linked to the account
+    transaction.begin()
+
+    user = UserCollection(client.app.session())\
+        .by_username('admin@example.org')
+
+    user.second_factor = {
+        'type': 'yubikey',
+        'data': 'ccccccbcgujh'
+    }
+
+    transaction.commit()
+
+    assert "nicht mit Ihrem Konto verknüpft" in sign(client, page, 'foobar')
+
+    # if the key doesn't work, the signature is not applied
+    with patch.object(Yubico, 'verify') as verify:
+        verify.return_value = False
+
+        assert "konnte nicht validiert werden" in sign(
+            client, page, 'ccccccbcgujhingjrdejhgfnuetrgigvejhhgbkugded')
+
+    assert not FileCollection(client.app.session()).query().one().signed
+
+    # once the signature has been applied, it can't be repeated
+    tape = module_path('onegov.org', 'tests/cassettes/ais-success.json')
+
+    with patch.object(Yubico, 'verify') as verify:
+        verify.return_value = True
+
+        with vcr.use_cassette(tape, record_mode='none'):
+
+            assert "signiert von admin@example.org" in sign(
+                client, page, 'ccccccbcgujhingjrdejhgfnuetrgigvejhhgbkugded')
+
+        with vcr.use_cassette(tape, record_mode='none'):
+
+            assert "Datei hat bereits eine digital Signatur" in sign(
+                client, page, 'ccccccbcgujhingjrdejhgfnuetrgigvejhhgbkugded')
+
+    # we should at this point see some useful metadata on the file
+    metadata = FileCollection(client.app.session())\
+        .query().one().signature_metadata
+
+    assert metadata['token'] == 'ccccccbcgujhingjrdejhgfnuetrgigvejhhgbkugded'
+    assert metadata['signee'] == 'admin@example.org'
+
+    # and we should see a message in the activity log
+    assert 'Datei signiert' in client.get('/timeline')